+++ src/utils/index.ts
@@ -21,12 +21,4 @@ export const safeParseInt = (str: string): number | null => {
   return value;
 };
 
-export const convertMillisecondsIntoDateValues = (milliseconds: number) => {
-  const days = Math.floor(milliseconds / (1000 * 60 * 60 * 24));
-  const hours = Math.floor((milliseconds % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
-  const minutes = Math.floor((milliseconds % (1000 * 60 * 60)) / (1000 * 60));
-  const seconds = Math.floor((milliseconds % (1000 * 60)) / 1000);
-  return { days, hours, minutes, seconds };
-};
-
 export const getScreenMaxWidthMediaQuery = (maxWidth: string) => `screen and (max-width: ${maxWidth})`;
